requirements.txt
faster-whisper
requests
nvidia-cublas-cu12
nvidia-cudnn-cu12
pyannote.audio
torchcodec
soundfile
torchaudio
pydub